+++ packages/oas-utils/src/entities/workspace/spec/index.ts
@@ -1,3 +1,4 @@
 export * from './parameters'
 export * from './requests'
+export * from './request-examples'
 export * from './components'
